display: Simplify profile matching
Previously a combination of EDID and display connector was used.
Now we only refer to the EDID of the display to determine if a profile
matches. In practice this means that if you plug the same display to a
different connector (e.g. HDMI-1 instead of HMDI-2) the profile will
still match while before this commit it didn't.